+++ docs/getting-started/quickstart/hello.py
@@ -6,7 +6,3 @@ app = App()
 @app.route("/")
 async def index(req, res):
     res.text = "Hello, world!"
-
-
-if __name__ == "__main__":
-    app.run()