+++ docs/guides/http/sse_example/app.py
@@ -38,7 +38,3 @@ class Events:
 @app.route("/")
 async def index(req, res):
     res.html = await templates.render("index.html")
-
-
-if __name__ == "__main__":
-    app.run()